Squash bugs in integration
Make sure to import into parent modules any children
Webpack will not do that for you automatically
