+++ client/src/app/SignUpForm.jsx
@@ -1,8 +1,9 @@
-class LoginForm extends React.Component {
+class SignUpForm extends React.Component {
   constructor(){
     super();
     this.state = {
       userName: '',
+      email: '',
       password: ''
     };
   }
@@ -12,6 +13,9 @@ class LoginForm extends React.Component {
     this.setState({userName: e.target.value});
   }
 
+  handleEmailChange(e) {
+    this.setState({email: e.target.value});
+  }
 
   handlePasswordChange(e) {
     this.setState({password: e.target.value});
@@ -31,6 +35,12 @@ class LoginForm extends React.Component {
         />
 
 
+        <input id="Email"
+          type="text" onChange={this.handleEmailChange.bind(this)}
+          value={this.state.email}
+          placeholder="Email"
+        />
+
         <input id="password"
           type="text" onChange={this.handlePasswordChange.bind(this)}
           value={this.state.password}
@@ -38,7 +48,7 @@ class LoginForm extends React.Component {
         />
 
         <button type="submit">
-          Login to DinnerParty
+          SignUp for DinnerParty
         </button>
 
       </form>
@@ -46,5 +56,4 @@ class LoginForm extends React.Component {
   }
 }
 
-
-window.LoginForm = LoginForm;
+export default SignUpForm;
