+++ client/src/index.html
@@ -7,17 +7,7 @@
     <!-- initial react app tag -->
     <div id="app"></div>
 
-    <!-- component tags -->
-    <script type="text/babel" src="app/GuestRow.js"></script>
-    <script type="text/babel" src="app/GuestsListTable.js"></script>
-    <script type="text/babel" src="app/CreatePartyForm.js"></script>
-    <script type="text/babel" src="app/PartyDetailsTable.js"></script>
-    <script type="text/babel" src="app/PartyRow.js"></script>
-    <script type="text/babel" src="app/SignUpForm.js"></script>
-    <script type="text/babel" src="app/LoginForm.js"></script>
-    <script type="text/babel" src="app/InviteFriendForm.js"></script>
-
     <!-- this is where the webpack.config serves the files, see output publicPath -->
-    <script src="bundle.js"></script>
+    <script type="text/javascript" src="bundle.js"></script>
   </body>
 </html>
